# Add get_psf_image options for performance-critical applications