These two options allow `get_psf_image` to run faster:
```
In [1]: from chandra_aca.aca_image import AcaPsfLibrary

In [2]: ap = AcaPsfLibrary()

In [3]: timeit ap.get_psf_image(100.05, 200.05)
82.7 µs ± 4.45 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [4]: timeit ap.get_psf_image(100.05, 200.05, aca_image=False)
37 µs ± 158 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [5]: timeit ap.get_psf_image(100.05, 200.05, aca_image=False, interpolation='nearest')
18.6 µs ± 100 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
If or when we put this into starcheck or star selection then a factor of 4 (plus downstream speedup by using `ndarray` instead of `ACAImage`) will be important.  There is also potential for numba-izing this, though at 18 usec this part might no longer be the bottleneck.