@taldcroft 2017-12-04
@jeanconn - in the battle between returning useless things in one case and returning different things in different cases, I took the latter path.  So I think this should be good to go.  I'm not in a huge hurry to release this, so it can wait until there is a need or another PR.